backend/requirements.txt
django
mysqlclient
pymysql
djangorestframework
djangorestframework-simplejwt
drf-yasg
torch
transformers
konlpy
sentence_transformers